+++ elasticsearch-20170613/src/Models/DescribeSnapshotSettingResponseBody/result.php
@@ -14,16 +14,25 @@ class result extends Model
     public $enable;
 
     /**
+     * @var string[]
+     */
+    public $indices;
+
+    /**
      * @var string
      */
     public $quartzRegex;
     protected $_name = [
         'enable' => 'Enable',
+        'indices' => 'Indices',
         'quartzRegex' => 'QuartzRegex',
     ];
 
     public function validate()
     {
+        if (\is_array($this->indices)) {
+            Model::validateArray($this->indices);
+        }
         parent::validate();
     }
 
@@ -32,6 +41,17 @@ class result extends Model
         $res = [];
         if (null !== $this->enable) {
             $res['Enable'] = $this->enable;
+        }
+
+        if (null !== $this->indices) {
+            if (\is_array($this->indices)) {
+                $res['Indices'] = [];
+                $n1 = 0;
+                foreach ($this->indices as $item1) {
+                    $res['Indices'][$n1] = $item1;
+                    ++$n1;
+                }
+            }
         }
 
         if (null !== $this->quartzRegex) {
@@ -51,6 +71,17 @@ class result extends Model
         $model = new self();
         if (isset($map['Enable'])) {
             $model->enable = $map['Enable'];
+        }
+
+        if (isset($map['Indices'])) {
+            if (!empty($map['Indices'])) {
+                $model->indices = [];
+                $n1 = 0;
+                foreach ($map['Indices'] as $item1) {
+                    $model->indices[$n1] = $item1;
+                    ++$n1;
+                }
+            }
         }
 
         if (isset($map['QuartzRegex'])) {