+++ elasticsearch-20170613/src/Models/ListDiagnosisItemsResponseBody/result.php
@@ -9,6 +9,16 @@ use AlibabaCloud\Dara\Model;
 class result extends Model
 {
     /**
+     * @var bool
+     */
+    public $billable;
+
+    /**
+     * @var string
+     */
+    public $category;
+
+    /**
      * @var string
      */
     public $description;
@@ -26,22 +36,53 @@ class result extends Model
     /**
      * @var string
      */
+    public $level;
+
+    /**
+     * @var string
+     */
     public $name;
+
+    /**
+     * @var int
+     */
+    public $sortOrder;
+
+    /**
+     * @var string[]
+     */
+    public $supportedModes;
     protected $_name = [
+        'billable' => 'billable',
+        'category' => 'category',
         'description' => 'description',
         'esApiRequired' => 'esApiRequired',
         'key' => 'key',
+        'level' => 'level',
         'name' => 'name',
+        'sortOrder' => 'sortOrder',
+        'supportedModes' => 'supportedModes',
     ];
 
     public function validate()
     {
+        if (\is_array($this->supportedModes)) {
+            Model::validateArray($this->supportedModes);
+        }
         parent::validate();
     }
 
     public function toArray($noStream = false)
     {
         $res = [];
+        if (null !== $this->billable) {
+            $res['billable'] = $this->billable;
+        }
+
+        if (null !== $this->category) {
+            $res['category'] = $this->category;
+        }
+
         if (null !== $this->description) {
             $res['description'] = $this->description;
         }
@@ -54,8 +95,27 @@ class result extends Model
             $res['key'] = $this->key;
         }
 
+        if (null !== $this->level) {
+            $res['level'] = $this->level;
+        }
+
         if (null !== $this->name) {
             $res['name'] = $this->name;
+        }
+
+        if (null !== $this->sortOrder) {
+            $res['sortOrder'] = $this->sortOrder;
+        }
+
+        if (null !== $this->supportedModes) {
+            if (\is_array($this->supportedModes)) {
+                $res['supportedModes'] = [];
+                $n1 = 0;
+                foreach ($this->supportedModes as $item1) {
+                    $res['supportedModes'][$n1] = $item1;
+                    ++$n1;
+                }
+            }
         }
 
         return $res;
@@ -69,6 +129,14 @@ class result extends Model
     public static function fromMap($map = [])
     {
         $model = new self();
+        if (isset($map['billable'])) {
+            $model->billable = $map['billable'];
+        }
+
+        if (isset($map['category'])) {
+            $model->category = $map['category'];
+        }
+
         if (isset($map['description'])) {
             $model->description = $map['description'];
         }
@@ -81,8 +149,27 @@ class result extends Model
             $model->key = $map['key'];
         }
 
+        if (isset($map['level'])) {
+            $model->level = $map['level'];
+        }
+
         if (isset($map['name'])) {
             $model->name = $map['name'];
+        }
+
+        if (isset($map['sortOrder'])) {
+            $model->sortOrder = $map['sortOrder'];
+        }
+
+        if (isset($map['supportedModes'])) {
+            if (!empty($map['supportedModes'])) {
+                $model->supportedModes = [];
+                $n1 = 0;
+                foreach ($map['supportedModes'] as $item1) {
+                    $model->supportedModes[$n1] = $item1;
+                    ++$n1;
+                }
+            }
         }
 
         return $model;